Renamed synonyms API namespaces
