+++ output/schema/validation-errors.json
@@ -1385,28 +1385,23 @@
         "response definition sql.delete_async:Response / body / instance_of - Non-leaf type cannot be used here: '_types:AcknowledgedResponseBase'"
       ]
     },
-    "synonym_rule.get": {
+    "synonyms.delete_synonym": {
       "request": [],
       "response": [
-        "interface definition synonyms._types:SynonymRule - Property 'id' is already defined in an ancestor class"
+        "response definition synonyms.delete_synonym:Response / body / instance_of - Non-leaf type cannot be used here: '_types:AcknowledgedResponseBase'"
       ]
     },
-    "synonyms.delete": {
-      "request": [],
-      "response": [
-        "response definition synonyms.delete:Response / body / instance_of - Non-leaf type cannot be used here: '_types:AcknowledgedResponseBase'"
-      ]
-    },
-    "synonyms.get": {
+    "synonyms.get_synonym": {
       "request": [],
       "response": [
         "interface definition synonyms._types:SynonymsSetUpdate / Property 'synonymRules' / array_of / instance_of - Non-leaf type cannot be used here: 'synonyms._types:SynonymRuleOptionalId'",
-        "interface definition synonyms._types:SynonymsSet - Property 'synonymRules' is already defined in an ancestor class"
+        "interface definition synonyms._types:SynonymsSet - Property 'synonymRules' is already defined in an ancestor class",
+        "interface definition synonyms._types:SynonymRule - Property 'id' is already defined in an ancestor class"
       ]
     },
-    "synonyms.put": {
+    "synonyms.put_synonym": {
       "request": [
-        "request definition synonyms.put:Request / body / instance_of - Non-leaf type cannot be used here: 'synonyms._types:SynonymsSetUpdate'"
+        "request definition synonyms.put_synonym:Request / body / instance_of - Non-leaf type cannot be used here: 'synonyms._types:SynonymsSetUpdate'"
       ],
       "response": []
     },
