+++ specification/_json_spec/synonyms.get_synonym.json
@@ -1,5 +1,5 @@
 {
-  "synonyms.get": {
+  "synonyms.get_synonym": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/get-synonyms-set.html",
       "description": "Retrieves a synonym set"
@@ -7,15 +7,19 @@
     "stability": "experimental",
     "visibility": "public",
     "headers": {
-      "accept": ["application/json"]
+      "accept": [
+        "application/json"
+      ]
     },
     "url": {
       "paths": [
         {
-          "path": "/_synonyms/{synonyms_set}",
-          "methods": ["GET"],
+          "path": "/_synonyms/{id}",
+          "methods": [
+            "GET"
+          ],
           "parts": {
-            "synonyms_set": {
+            "id": {
               "type": "string",
               "description": "The name of the synonyms set to be retrieved"
             }
