+++ specification/_json_spec/synonyms.get_synonym_rule.json
@@ -1,5 +1,5 @@
 {
-  "synonym_rule.get": {
+  "synonyms.get_synonym_rule": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/get-synonym-rule.html",
       "description": "Retrieves a synonym rule from a synonym set"
@@ -7,20 +7,26 @@
     "stability": "experimental",
     "visibility": "public",
     "headers": {
-      "accept": ["application/json"],
-      "content_type": ["application/json"]
+      "accept": [
+        "application/json"
+      ],
+      "content_type": [
+        "application/json"
+      ]
     },
     "url": {
       "paths": [
         {
-          "path": "/_synonyms/{synonyms_set}/{synonym_rule}",
-          "methods": ["GET"],
+          "path": "/_synonyms/{set_id}/{rule_id}",
+          "methods": [
+            "GET"
+          ],
           "parts": {
-            "synonyms_set": {
+            "set_id": {
               "type": "string",
               "description": "The id of the synonym set to retrieve the synonym rule from"
             },
-            "synonym_rule": {
+            "rule_id": {
               "type": "string",
               "description": "The id of the synonym rule to retrieve"
             }
