+++ specification/_json_spec/synonyms.put_synonym.json
@@ -1,5 +1,5 @@
 {
-  "synonyms.put": {
+  "synonyms.put_synonym": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/put-synonyms-set.html",
       "description": "Creates or updates a synonyms set"
@@ -7,18 +7,24 @@
     "stability": "experimental",
     "visibility": "public",
     "headers": {
-      "accept": ["application/json"],
-      "content_type": ["application/json"]
+      "accept": [
+        "application/json"
+      ],
+      "content_type": [
+        "application/json"
+      ]
     },
     "url": {
       "paths": [
         {
-          "path": "/_synonyms/{synonyms_set}",
-          "methods": ["PUT"],
+          "path": "/_synonyms/{id}",
+          "methods": [
+            "PUT"
+          ],
           "parts": {
-            "synonyms_set": {
+            "id": {
               "type": "string",
-              "description": "The name of the synonyms set to be created or updated"
+              "description": "The id of the synonyms set to be created or updated"
             }
           }
         }
