+++ specification/synonyms/delete_synonym/SynonymsDeleteResponse.ts
