+++ specification/synonyms/delete_synonym_rule/SynonymRuleDeleteResponse.ts
@@ -16,25 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { RequestBase } from '@_types/Base'
-import { Name } from '@_types/common'
-import { integer } from '@_types/Numeric'
 
-/**
- * Retrieves a summary of all defined synonym sets
- * @rest_spec_name synonyms_sets.get
- * @availability stack since=8.10.0 stability=experimental
- * @availability serverless stability=experimental visibility=public
- */
-interface Request extends RequestBase {
-  query_parameters: {
-    /**
-     * Starting offset (default: 0)
-     */
-    from?: integer
-    /**
-     * specifies a max number of results to get (default: 10)
-     */
-    size?: integer
-  }
+import { SynonymsUpdateResult } from '../_types/SynonymsUpdateResult'
+
+export class Response {
+  body: SynonymsUpdateResult
 }
