+++ specification/synonyms/get_synonym/SynonymsGetResponse.ts
