+++ specification/synonyms/get_synonyms_sets/SynonymsSetsGetResponse.ts
