+++ specification/synonyms/put_synonym/SynonymsPutResponse.ts
